requirements.txt
streamlit>=1.28.0
requests>=2.31.0
fastapi>=0.104.0
uvicorn>=0.24.0
pandas>=2.0.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
